added clustering util
Should be usable - on occasion, when you have all user ids and embeddings available, call into this utility to generate clusters.  Overall model should try to load these files if possible, and use them with generate_user_embedding_vector() to create the combined cluster-user representations for the actual document retrieval.
If it seems as if results are all too similar, either adjusting the beta passed to generate or the clustering technique to align better with the actual data structure (may be beyond scope) to cause algo to spread results a bit more.
This right now is a simple implementation of the clustering algorithm, more of a proof of concept than a polished capability.